SearchResultsViewControlleにファイルを分割
